re-organize content
