+++ package.json
@@ -9,14 +9,14 @@
     "node": ">=18.x"
   },
   "scripts": {
-    "start": "node app.js",
-    "register": "node commands.js",
-    "dev": "nodemon app.js"
+    "start": "node src/app.js",
+    "register": "node src/register.js",
+    "dev": "nodemon src/app.js"
   },
   "author": "Shay DeWael",
   "license": "MIT",
   "dependencies": {
-    "discord-interactions": "^4.0.0",
+    "discord-interactions": "^4.1.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.2"
   },
